LPD-98752 Move the CMS site initializer util to the API module
